+++ packages/forge-angular/src/dev/app/views/examples/table-example/types.ts
@@ -1,0 +1,12 @@
+export interface IJournal {
+  number: number;
+  year: number;
+  month: number;
+  source: string;
+  status: number;
+  enteredBy: string;
+  description: string;
+  debits: number;
+  credits: number;
+  attachments: boolean;
+}