+++ packages/forge-angular/src/dev/app/views/home/home.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { DemoCardComponent } from '../../components/demo-card/demo-card.component';
+
+@Component({
+  selector: 'app-home',
+  templateUrl: './home.component.html',
+  styleUrls: ['./home.component.scss'],
+  imports: [DemoCardComponent]
+})
+export class HomeComponent {}