+++ packages/forge-angular/src/lib/app-bar-profile-button/index.ts
@@ -1,0 +1,2 @@
+export * from './app-bar-profile-button.component';
+export * from './app-bar-profile-button.module';