+++ packages/forge-angular/src/lib/open-icon/index.ts
@@ -1,0 +1,2 @@
+export * from './open-icon.component';
+export * from './open-icon.module';