+++ packages/forge-angular/src/lib/overlay/index.ts
@@ -1,0 +1,2 @@
+export * from './overlay.component';
+export * from './overlay.module';