+++ packages/forge-angular/src/lib/timestamp/index.ts
@@ -1,0 +1,2 @@
+export * from './timestamp.component';
+export * from './timestamp.module';